Task 1-Number Guessing Game, Task 2-ATM Interface
The first task is to create a number guessing game that allows user to guess a random number chosen from 1-100. It also gives then hints to try guessing a larger or a small number. It also gives points to the user according to the number of chances taken to guess.
The second task is to create ATM Interface that allows user to add, delete, access accounts, and showing list of accounts and to finally exit from the system.